Moved rete builder methods to rete_build.py
